Reviews progress messages output.
It is not pushing a website to GitHub, but a branch that will (likely)
be used as the source of files for a website.

Also says the branch name being pushed and also in case the command
fails, the branch name will be reported in the error annotations of the
action.