Re-creates working directory if git-rm removes it.
If an existing branch is to be overwritten and the documentation should
be deployed to a subdirectory of that branch, when wiping with git-rm,
it will remove also the current working directory. Simply re-creating it
should suffice to keep the flow.

In case the directory is the only entry in an upper level directory,
git-rm will also remove directories up to root while they meet this
criteria (no empty directories left in the tree). Thus, the mkdir
command should recursively create upper level directories as needed (-p
flag).

Notice GNU mkdir (from coreutils) accepts friendlier --parents argument,
whereas OS X uses BSD version, which claims to be IEEE Std 10003.2
(POSIX.2) compliant. As a result, only the -p flag is accepted across
platforms.

Additionally, at least for Darwin's apfs and linux ext filesystems, when the
directory is removed and re-created, we need to re-chdir into it because
likely it now has a new logical filesystem entry.

Related GitHub issue: #2.