Refactors config.js' ID handlers.
Now instead of regex matching the ID, the file is parsed by node and the
topmost ID property value is extracted from the object.

When replacing the ID, the part until the first curly bracket ({),
hopefully 'var __DOCS_CONFIG__ = ' is extracted from the file and
concatenated with the re-stringified value after the file is interpreted
and topmost ID reassigned from a node-parsed script.

Related GitHub issue: retypeapp/action-github-pages#6.