Adds tags to all yml samples' 'uses:' occurrences.
Not having an explicit tag would mean it should pull the default branch
which, for some reason (according to retypeapp/retype#248) GitHub
actions didn't accept the `owner/repo` path without '@ref' instead of
choosing default branch (which should work just like '@v1' would).

Nevertheless it was never supposed to be documented the way it is and we
have long conventioned to use the '@v1' tag, so with this the
documentation is also conforming to our (and GitHub actions') standards.