Fixes possible git-checkout ambiguity.
When checking out the specified branch, if the default branch has a file
matching the branch name, git-checkout won't be sure what to do if just
'git checkout <branch>' is executed. So now we specify the branch name
with '-b' -and- also instruct that the branch should be tracking its
remote counterpart.

Related GitHub issue: retypeapp/action-github-pages#13